+++ boxoffice/extapi/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from .razorpay import *
+from .razorpay_status import *
